backend/requirements.txt
asgiref==3.5.2
astroid==2.12.5
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
dill==0.3.5.1
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
flake8==5.0.4
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.15.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pylint==2.15.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
